[ALPS08520907] GPU: GPUCORE-39981 for CVE-2023-5427
[Description]
The fix for GPUCORE-36524 introduced a use after free vulnerability on
kernel side. Due to the race between poll and close system calls,
issued on /dev/mali0 file, kernel could end up accessing the event
waitqueue of Kbase context after the memory for context was freed on
the close of file.
The event waitqueue of Kbase context is added to a list, maintained by
the kernel, from the poll file operation method of /dev/mali0 file.
At the same time, kernel also takes a reference on the mali file.
The waitqueue is removed from the list by kernel and the reference is
dropped just before the end of poll system call, i.e. after the poll
file operation method has completed.
Prior to GPUCORE-36524, there was no issue as lifetime of Kbase context
matched that of the mali file instance. But after GPUCORE-36524, Kbase
context can get destroyed & its memory freed before the last reference
on mali file is dropped. Thus it is possible that the memory for context
can get freed as soon as the poll file operation method completes.

To fix the issue the event waitqueue is moved to struct kbase_file.
A defect test is also added to exercise the use after free scenario.

[Patch]
(c6096/6) GPUCORE-39981: Fix UAF for the event waitqueue of Kbase context

[ARM Ticket]
03451829

MTK-Commit-Id: 25a78a019630792c0938b2e72ae68527e0468009

Signed-off-by: Daniel Tao <daniel.tao@mediatek.com>
CR-Id: ALPS08520907
Feature: ARM GPU DDK driver
Change-Id: Ib27d344761c248c104ef77cc1fe6a701150a2748