[ALPS08684229] GPU: GPUCORE-40918 for CVE-2023-6363
[Description]
This patchset defers the registration of a KCPU queue to the KCPU
context until after both the queue and the metadata are completely set
up. This fixes a use-after free, where this dangling pointer is left in
the context. To simplify the control flow, the initialization of the
metadata object has been moved to a separate function.

[Patch]
(c6332/1) CVE-2023-6363 GPUCORE-40918 Defer registration of new KCPU queues

[ARM Ticket]
03460114

MTK-Commit-Id: e4a0420bd20b430457665a81c2480997b032810a

Signed-off-by: Daniel Tao <daniel.tao@mediatek.com>
CR-Id: ALPS08684229
Feature: ARM GPU DDK driver
Change-Id: I3849721e5811d15e07e1f9206b416ab07e57d6bb